refactor: domain, api 타입 폴더 분리
