+++ src/features/movie/api/getMovieDetail.ts
@@ -1,5 +1,5 @@
 import { mapToMovieDetail } from "../../../shared/domain/mapToMovie";
-import { MovieDetail } from "../../../shared/types/movies";
+import { MovieDetail } from "../../../shared/types/domain/movies";
 import { createApiUrl, fetchApi } from "../../../shared/utils/apiUtils";
 
 const getMovieDetail = async (id: number): Promise<MovieDetail> => {
